backend/requirements.txt
flask==2.3.3
flask-cors==6.0.0
python-dotenv==1.0.0
langchain==0.2.5
openai==0.27.8
python-multipart==0.0.18
pdfminer.six==20221105
docx2txt==0.8
